services/redis-upstash/pyproject.toml
[project]
name = "redis-upstash"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv==1.2.1",
    "redis==7.2.1",
    "upstash-redis==1.6.0",
]
